refactor: put ps2 into own module
