+++ src/ps2/controller.rs
@@ -1,6 +1,8 @@
 /// https://wiki.osdev.org/%228042%22_PS/2_Controller
 use core::arch::asm;
 
+use super::{Key, scancodes::SCANCODE_TO_KEY};
+
 const DATA_PORT: u16 = 0x60;
 const STATUS_PORT: u16 = 0x64;
 const COMMAND_PORT: u16 = 0x64;
@@ -413,351 +415,3 @@ unsafe fn write(port: u16, val: u8) {
         );
     }
 }
-
-#[repr(u8)]
-#[allow(unused)]
-#[derive(Clone, Copy, PartialEq, PartialOrd)]
-pub enum Key {
-    Escape,
-    Tab,
-    Enter,
-    ArrowUp,
-    Backspace,
-    ArrowDown,
-    ArrowLeft,
-    ArrowRight,
-    LeftAlt,
-    LeftShift,
-    RightAlt,
-    RightShift,
-    LeftCtrl,
-    RightCtrl,
-    CapsLock,
-    NumLock,
-    ScrollLock,
-    F1,
-    F2,
-    F3,
-    F4,
-    F5,
-    F6,
-    F7,
-    F8,
-    F9,
-    F10,
-    F11,
-    F12,
-    A = b'a',
-    B = b'b',
-    C = b'c',
-    D = b'd',
-    E = b'e',
-    F = b'f',
-    G = b'g',
-    H = b'h',
-    I = b'i',
-    J = b'j',
-    K = b'k',
-    L = b'l',
-    M = b'm',
-    N = b'n',
-    O = b'o',
-    P = b'p',
-    Q = b'q',
-    R = b'r',
-    S = b's',
-    T = b't',
-    U = b'u',
-    V = b'v',
-    W = b'w',
-    X = b'x',
-    Y = b'y',
-    Z = b'z',
-    N0 = b'0',
-    N1 = b'1',
-    N2 = b'2',
-    N3 = b'3',
-    N4 = b'4',
-    N5 = b'5',
-    N6 = b'6',
-    N7 = b'7',
-    N8 = b'8',
-    N9 = b'9',
-    Point = b'.',
-    Space = b' ',
-    Equal = b'=',
-    Comma = b',',
-    Backtick = b'`',
-    Semicolon = b';',
-    Backslash = b'\\',
-    SingleQuote = b'\'',
-    SquareBracketsOpen = b'[',
-    SquareBracketsClosed = b']',
-    Add = b'+',
-    Sub = b'-',
-    Mul = b'*',
-    Div = b'/',
-}
-
-use Key::*;
-/// Conversion table for all characters currently supported by our kernel for PS2 input.
-///
-/// https://wiki.osdev.org/PS/2_Keyboard#Scan_Code_Set_2
-const SCANCODE_TO_KEY: [(u8, Option<Key>); 256] = [
-    (0x00, None),
-    (0x01, Some(F9)),
-    (0x02, None),
-    (0x03, Some(F5)),
-    (0x04, Some(F4)),
-    (0x05, Some(F1)),
-    (0x06, Some(F2)),
-    (0x07, Some(F12)),
-    (0x08, None),
-    (0x09, Some(F10)),
-    (0x0A, Some(F8)),
-    (0x0B, Some(F6)),
-    (0x0C, Some(F8)),
-    (0x0D, Some(Tab)),
-    (0x0E, Some(Backtick)),
-    (0x0F, None),
-    (0x10, None),
-    (0x11, Some(LeftAlt)),
-    (0x12, Some(LeftShift)),
-    (0x13, None),
-    (0x14, Some(LeftCtrl)),
-    (0x15, Some(Q)),
-    (0x16, Some(N1)),
-    (0x17, None),
-    (0x18, None),
-    (0x19, None),
-    (0x1A, Some(Z)),
-    (0x1B, Some(S)),
-    (0x1C, Some(A)),
-    (0x1D, Some(W)),
-    (0x1E, Some(N2)),
-    (0x1F, None),
-    (0x20, None),
-    (0x21, Some(C)),
-    (0x22, Some(X)),
-    (0x23, Some(D)),
-    (0x24, Some(E)),
-    (0x25, Some(N4)),
-    (0x26, Some(N3)),
-    (0x27, None),
-    (0x28, None),
-    (0x29, Some(Space)),
-    (0x2A, Some(V)),
-    (0x2B, Some(F)),
-    (0x2C, Some(T)),
-    (0x2D, Some(R)),
-    (0x2E, Some(N5)),
-    (0x2F, None),
-    (0x30, None),
-    (0x31, Some(N)),
-    (0x32, Some(B)),
-    (0x33, Some(H)),
-    (0x34, Some(G)),
-    (0x35, Some(Y)),
-    (0x36, Some(N6)),
-    (0x37, None),
-    (0x38, None),
-    (0x39, None),
-    (0x3A, Some(M)),
-    (0x3B, Some(J)),
-    (0x3C, Some(U)),
-    (0x3D, Some(N7)),
-    (0x3E, Some(N8)),
-    (0x3F, None),
-    (0x40, None),
-    (0x41, Some(Comma)),
-    (0x42, Some(K)),
-    (0x43, Some(I)),
-    (0x44, Some(O)),
-    (0x45, Some(N0)),
-    (0x46, Some(N9)),
-    (0x47, None),
-    (0x48, None),
-    (0x49, Some(Point)),
-    (0x4A, Some(Div)),
-    (0x4B, Some(L)),
-    (0x4C, Some(Semicolon)),
-    (0x4D, Some(P)),
-    (0x4E, Some(Sub)),
-    (0x4F, None),
-    (0x50, None),
-    (0x51, None),
-    (0x52, Some(SingleQuote)),
-    (0x53, None),
-    (0x54, Some(SquareBracketsOpen)),
-    (0x55, Some(Equal)),
-    (0x56, None),
-    (0x57, None),
-    (0x58, Some(CapsLock)),
-    (0x59, Some(RightShift)),
-    (0x5A, Some(Enter)),
-    (0x5B, Some(SquareBracketsClosed)),
-    (0x5C, None),
-    (0x5D, Some(Backslash)),
-    (0x5E, None),
-    (0x5F, None),
-    (0x60, None),
-    (0x61, None),
-    (0x62, None),
-    (0x63, None),
-    (0x64, None),
-    (0x65, None),
-    (0x66, Some(Backspace)),
-    (0x67, None),
-    (0x68, None),
-    (0x69, Some(N1)),
-    (0x6A, None),
-    (0x6B, Some(N4)),
-    (0x6C, Some(N7)),
-    (0x6D, None),
-    (0x6E, None),
-    (0x6F, None),
-    (0x70, Some(N0)),
-    (0x71, Some(Point)),
-    (0x72, Some(N2)),
-    (0x73, Some(N5)),
-    (0x74, Some(N6)),
-    (0x75, Some(N8)),
-    (0x76, Some(Escape)),
-    (0x77, Some(NumLock)),
-    (0x78, Some(F11)),
-    (0x79, Some(Add)),
-    (0x7A, Some(N3)),
-    (0x7B, Some(Sub)),
-    (0x7C, Some(Mul)),
-    (0x7D, Some(N9)),
-    (0x7E, Some(ScrollLock)),
-    (0x7F, None),
-    (0x80, None),
-    (0x81, None),
-    (0x82, None),
-    (0x83, Some(F7)),
-    (0x84, None),
-    (0x85, None),
-    (0x86, None),
-    (0x87, None),
-    (0x88, None),
-    (0x89, None),
-    (0x8A, None),
-    (0x8B, None),
-    (0x8C, None),
-    (0x8D, None),
-    (0x8E, None),
-    (0x8F, None),
-    (0x90, None),
-    (0x91, None),
-    (0x92, None),
-    (0x93, None),
-    (0x94, None),
-    (0x95, None),
-    (0x96, None),
-    (0x97, None),
-    (0x98, None),
-    (0x99, None),
-    (0x9A, None),
-    (0x9B, None),
-    (0x9C, None),
-    (0x9D, None),
-    (0x9E, None),
-    (0x9F, None),
-    (0xA0, None),
-    (0xA1, None),
-    (0xA2, None),
-    (0xA3, None),
-    (0xA4, None),
-    (0xA5, None),
-    (0xA6, None),
-    (0xA7, None),
-    (0xA8, None),
-    (0xA9, None),
-    (0xAA, None),
-    (0xAB, None),
-    (0xAC, None),
-    (0xAD, None),
-    (0xAE, None),
-    (0xAF, None),
-    (0xB0, None),
-    (0xB1, None),
-    (0xB2, None),
-    (0xB3, None),
-    (0xB4, None),
-    (0xB5, None),
-    (0xB6, None),
-    (0xB7, None),
-    (0xB8, None),
-    (0xB9, None),
-    (0xBA, None),
-    (0xBB, None),
-    (0xBC, None),
-    (0xBD, None),
-    (0xBE, None),
-    (0xBF, None),
-    (0xC0, None),
-    (0xC1, None),
-    (0xC2, None),
-    (0xC3, None),
-    (0xC4, None),
-    (0xC5, None),
-    (0xC6, None),
-    (0xC7, None),
-    (0xC8, None),
-    (0xC9, None),
-    (0xCA, None),
-    (0xCB, None),
-    (0xCC, None),
-    (0xCD, None),
-    (0xCE, None),
-    (0xCF, None),
-    (0xD0, None),
-    (0xD1, None),
-    (0xD2, None),
-    (0xD3, None),
-    (0xD4, None),
-    (0xD5, None),
-    (0xD6, None),
-    (0xD7, None),
-    (0xD8, None),
-    (0xD9, None),
-    (0xDA, None),
-    (0xDB, None),
-    (0xDC, None),
-    (0xDD, None),
-    (0xDE, None),
-    (0xDF, None),
-    (0xE0, None),
-    (0xE1, None),
-    (0xE2, None),
-    (0xE3, None),
-    (0xE4, None),
-    (0xE5, None),
-    (0xE6, None),
-    (0xE7, None),
-    (0xE8, None),
-    (0xE9, None),
-    (0xEA, None),
-    (0xEB, None),
-    (0xEC, None),
-    (0xED, None),
-    (0xEE, None),
-    (0xEF, None),
-    (0xF0, None),
-    (0xF1, None),
-    (0xF2, None),
-    (0xF3, None),
-    (0xF4, None),
-    (0xF5, None),
-    (0xF6, None),
-    (0xF7, None),
-    (0xF8, None),
-    (0xF9, None),
-    (0xFA, None),
-    (0xFB, None),
-    (0xFC, None),
-    (0xFD, None),
-    (0xFE, None),
-    (0xFF, None),
-];
