+++ src/ps2/mod.rs
@@ -1,0 +1,5 @@
+mod controller;
+mod scancodes;
+
+pub use controller::{init, read_if_ready};
+pub use scancodes::Key;
