+++ src/terminal/mod.rs
@@ -1,5 +1,4 @@
 pub mod cursor;
-pub mod ps2;
 mod screen;
 #[allow(clippy::module_inception)]
 pub mod terminal;
